Refactored some code to improve some dependencies and started using PoolWatchs to display some information
