+++ node/actors/network/Cargo.toml
@@ -26,6 +26,14 @@ thiserror.workspace = true
 tracing.workspace = true
 vise.workspace = true
 tokio.workspace = true
+tokio-rustls.workspace = true
+hyper.workspace = true
+http-body-util.workspace = true
+hyper-util.workspace = true
+tls-listener.workspace = true
+rustls-pemfile.workspace = true
+base64.workspace = true
+build_html.workspace = true
 
 [dev-dependencies]
 assert_matches.workspace = true
