+++ node/actors/network/src/lib.rs
@@ -13,6 +13,7 @@ mod config;
 pub mod consensus;
 mod frame;
 pub mod gossip;
+pub mod http;
 pub mod io;
 mod metrics;
 mod mux;
