+++ node/libs/utils/Cargo.toml
@@ -8,7 +8,6 @@ license.workspace = true
 
 [dependencies]
 zksync_concurrency.workspace = true
-
 rand.workspace = true
 thiserror.workspace = true
